Add the delete completed task logic.
