+++ src/updateTask.js
@@ -28,3 +28,14 @@ export function updateTask() {
     });
   });
 }
+
+export function deleteTask() {
+  const clearTask = document.querySelector('.clear-task');
+  clearTask.addEventListener('click', () => {
+    if (getTask().length > 0) {
+      const filterCompliteTask = getTask().filter((task) => task.completed !== true);
+      localStorage.setItem('Task-list', JSON.stringify(filterCompliteTask));
+      location.reload();
+    }
+  });
+}
